moved src/win32/debug.c contents into src/win32.c:
The 'lite' packages lacked that file, so MSVC builds with DEBUG defined would fail.
Fixes: https://github.com/libxmp/libxmp/issues/420
